devop: add "the-wrapper-wallet" to swap layout
